+++ apps/backend/tests/unit/organisation-admin.routes.test.ts
@@ -2,7 +2,7 @@ import type { NextFunction, Request, Response } from 'express';
 import request from 'supertest';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { createApp } from '../../src/app.js';
-import { clearOrganisationAdminRateLimitStores } from '../../src/middleware/organisationAdminRateLimit.js';
+import { clearOrganisationAdminRateLimitStores } from '../../src/routes/organisation-admin.routes.js';
 
 const actorUserId = '33333333-3333-4333-8333-333333333333';
 const organisationId = '11111111-1111-4111-8111-111111111111';
